Properly handle the GET calls to the API on ProductActivity
The way these calls were being handled was ugly and not efficient. Firing a second GET call on the listener of the first one is not the best way to do it, plus the handling of the loading dialog was not OK.
Instead let's create a new AsyncTask to handle both calls at the same time, and encapsulate the Loading dialog inside this class.